+++ src/Notification.php
@@ -9,20 +9,26 @@ class Notification
     public function __construct(
         protected string $message,
         protected ?string $title,
-        protected ?string $type = null
+        protected ?string $type = null,
+        protected bool $sanitize = true,
     ) {
-        $this->message = htmlspecialchars($message, ENT_QUOTES);
-        $this->title = $title ? htmlspecialchars($title, ENT_QUOTES) : '';
         $this->type = $type ?? NotificationType::$info;
     }
 
     protected function asArray(): array
     {
-        return [
-            'message' => $this->message,
-            'title' => $this->title ?? '',
-            'type' => $this->type,
-        ];
+        $message = $this->sanitize ? htmlspecialchars($this->message, ENT_QUOTES) : $this->message;
+        $title = $this->sanitize && $this->title ? htmlspecialchars($this->title, ENT_QUOTES) : $this->title;
+        $type = $this->type;
+
+        return compact('message', 'title', 'type');
+    }
+
+    public function doNotSanitize(): Notification
+    {
+        $this->sanitize = false;
+
+        return $this;
     }
 
     public static function make(